add image blank migration, type annotation in validate file, checking if image exist in signal

+++ airport/signals.py
@@ -8,5 +8,5 @@ from airport.models import Airplane
 
 @receiver(pre_delete, sender=Airplane)
 def delete_image(sender, instance, **kwargs):
-    if os.path.isfile(instance.image.path):
+    if instance.image and os.path.isfile(instance.image.path):
         os.remove(instance.image.path)
